Move diagramming controller connection from content container to lesson show view
Was originally in the content container component to be shared between
the lesson view and markdown preview tool, but the markdown preview tool
needs a different connection/action structure to render mermaid diagrams
properly.

Moving it out of the content container means the diagramming#connect method
can be essentially returned to original using the default .mermaid selector.