Feature: Provide loading feedback when completing a lesson on a slow connection (#4942)
Because:
- The complete button will appear to "hang" on slower connections.

This commit:
- Adds loading spinners to the complete button on lesson and course
pages, that will only appear when toggling complete and the connection
is slow.
- Bumps turbo-rails to latest patch version so we can use the :morph
option when replacing content with turbo streams and avoid hover flashes
on buttons as the content is replaced.
- Switch the complete button on course pages from links to buttons as
they are not making GET requests.
- Removes some old Stimulus JS hooks for a complete controller that has
been removed long ago.

Lesson page loading spinner:


https://github.com/user-attachments/assets/73ece768-9f82-4003-9bd9-5ed1260a5f7c

Course page loading spinner:


https://github.com/user-attachments/assets/4e066ae8-81ed-4632-93c0-b0b666c205fc