Feature: Center lesson content (#5144)
Because:
- Previously, to keep the lesson content and sidebar within the width of
our page container, which was dictated by the navbar, we positioned
lesson content to the left instead of center on the page. This gave us
room within the 7xl width to contain both the content and the sidebar.
- But it looked really weird when you got the bottom of the page when
completing a lesson, where content was again centered.

This commit:
- Expand the navbar to full width so we no longer have to confine page
content within it's width.
- Use a 3 column grid for lesson content - the left column will be
empty, the middle column will make sure lesson content is dead center,
and the right column will contain the lesson toc sidebar.
- Move the lesson header within the same grid so it can stay in sync
with lesson content responsive movements.