Chore: Refactor controller feat flag gates (#5164)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
The current checks for `Flipper` need to be repeated across any
controller action that requires a feature flag check. There are a couple
of compelling reasons to DRY this up:
1. It requires more work than necessary to remove the feature flag
checks from a controller as you have to go and edit each action in the
controller--and that's also going to be more error prone than a
potentially simple one-line edit.
2. If you want to do something like change where the redirect goes when
the feature is disabled, you again have to edit every place `Flipper` is
checked.

I also just think there are benefits to establishing a more Rails-y
convention on checking feature flags in the controller.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->
- Extracts a concern that wraps the `Flipper` check
- `extend` the concern into `ApplicationController`
- Refactor the `InterviewSurveyController` to use the new concern

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->
Closes #XXXXX

## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->
This is pretty shamelessly stolen from
https://kaspth.com/posts/clear-up-your-rails-controllers-with-before_action-wrappers

The article argues for a different redirect when a feature is disabled,
but I just kept what was there previously. Not sure it matters a ton in
our context. Anything that tries to be opaque about what the user is
accessing probably doesn't have much value with a fully open source app.
Willing to change it though!

## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
    - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
    - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [x] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [x] If applicable, this PR includes new or updated automated tests