Fix linting errors in application.tailwind.css
Now caught by stylelint's new path.
`import-notation: url` is the default, but url() in @import is optional.
Appeasing the linter for this would make things inconsistent with paths
in at-rules like @config, which stylelint won't be able to lint for
url(), so it's more sensible to configure it for `string`.